build(deps): bump tree-sitter to 0.27
tree-sitter 0.27 ties `Node::kind()` to the tree's lifetime, since a wasm
language can be freed. The node tracker and captured nodes key on the
kind as `&'static str` to avoid an allocation per node, so tracker-bound
call sites now go through `static_node_kind`, whose lifetime widening is
sound because every `Language` kakehashi parses with is compiled in or
comes from an immortal (never dlclose'd) library, and `wasm` is off. Kinds
and field names that only reach a JSON response are copied instead.

The other breaks are mechanical: `QueryMatch::captures` is now a method,
and `Node::child_count` returns `u32`. ABI support is unchanged (13-15),
so installed parsers keep loading.

+++ src/document/model.rs
@@ -505,7 +505,7 @@ mod tests {
         let close = root
             .named_child(0)
             .and_then(|function| function.child_by_field_name("body"))
-            .and_then(|body| body.child(body.child_count().saturating_sub(1) as u32));
+            .and_then(|body| body.child(body.child_count().saturating_sub(1)));
         root.end_byte() == text.len()
             && close.is_some_and(|close| {
                 close.kind() == "}" && Some(close.start_byte()) == text.rfind('}')
